chore(node): move `MinimalPluginContext` to a standalone file
