+++ packages/rolldown/src/index.ts
@@ -49,7 +49,7 @@ import { RenderedChunk, PreRenderedChunk } from './binding'
 import { PartialNull } from './types/utils'
 import { NormalizedInputOptions } from './options/normalized-input-options'
 import { ModuleInfo } from './types/module-info'
-import { MinimalPluginContext } from './log/logger'
+import { MinimalPluginContext } from './plugin/minimal-plugin-context'
 import { ExistingRawSourceMap, SourceMapInput } from './types/sourcemap'
 import { OutputBundle } from './types/output-bundle'
 import { version } from '../package.json'
