+++ packages/rolldown/src/log/logger.ts
@@ -15,62 +15,12 @@ import {
   type LogLevelOption,
   type LogLevel,
 } from './logging'
-import { error, logPluginError } from './logs'
-import { getLogHandler, normalizeLog } from './logHandler'
+import { error } from './logs'
+import { normalizeLog } from './logHandler'
 import type { InputOptions } from '../options/input-options'
-import type { NormalizedInputOptions } from '../options/normalized-input-options'
 import path from 'node:path'
 import { VERSION } from '..'
 import { getSortedPlugins } from '../plugin/plugin-driver'
-
-export interface PluginContextMeta {
-  rollupVersion: string
-  rolldownVersion: string
-  watchMode: boolean
-}
-
-export class MinimalPluginContext {
-  info: LoggingFunction
-  warn: LoggingFunction
-  debug: LoggingFunction
-  meta: PluginContextMeta
-  readonly error: (error: RollupError | string) => never
-
-  constructor(options: NormalizedInputOptions, plugin: Plugin) {
-    const onLog = options.onLog as LogHandler
-    const pluginName = plugin.name || 'unknown'
-    const logLevel = options.logLevel || LOG_LEVEL_INFO
-    this.debug = getLogHandler(
-      LOG_LEVEL_DEBUG,
-      'PLUGIN_LOG',
-      onLog,
-      pluginName,
-      logLevel,
-    )
-    this.info = getLogHandler(
-      LOG_LEVEL_INFO,
-      'PLUGIN_LOG',
-      onLog,
-      pluginName,
-      logLevel,
-    )
-    this.warn = getLogHandler(
-      LOG_LEVEL_WARN,
-      'PLUGIN_WARNING',
-      onLog,
-      pluginName,
-      logLevel,
-    )
-    this.error = (e): never => {
-      return error(logPluginError(normalizeLog(e), pluginName))
-    }
-    this.meta = {
-      rollupVersion: '4.23.0',
-      rolldownVersion: VERSION,
-      watchMode: false,
-    }
-  }
-}
 
 export function getLogger(
   plugins: Plugin[],
