+++ packages/rolldown/src/plugin/index.ts
@@ -18,7 +18,7 @@ import type { TransformPluginContext } from './transform-plugin-context'
 import type { NormalizedOutputOptions } from '../options/normalized-output-options'
 import type { LogLevel } from '../log/logging'
 import type { RollupLog } from '../rollup'
-import type { MinimalPluginContext } from '../log/logger'
+import type { MinimalPluginContext } from './minimal-plugin-context'
 import { InputOptions, OutputOptions } from '..'
 import { BuiltinPlugin } from '../builtin-plugin/constructors'
 import { ParallelPlugin } from './parallel-plugin'
